feat: add unit test and remove useless tests
