+++ pyproject.toml
@@ -340,3 +340,7 @@ packages = ["src/solace_ai_connector"]
 [tool.ruff]
 lint.select = ["E4", "E7", "E9", "F"]
 lint.ignore = ["F401", "E731"]
+
+[tool.pytest.ini_options]
+# Register pytest-mock plugin explicitly
+addopts = "-p pytest_mock"
